Fix Fischertecknik update functions
Fischertecknik hardware returns the input status after you send the output update message. Therefore, it is not possible to just read the inputs and then update the outputs. My quick solution for that is to update both inputs and outputs twice on the cycle, one on the updateBuffersIn() and the other on the updateBuffersOut(). This of course double the time required for I/O handling, but at least guarantees that the scan cycle starts with the accurate representation of the inputs and ends with the accurate representation of the outputs